--- module/psu/src/mod_psu_device_api_private.h
@@ -1,16 +1,0 @@
-/*
- * Arm SCP/MCP Software
- * Copyright (c) 2017-2019, Arm Limited and Contributors. All rights reserved.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-#ifndef MOD_PSU_DEVICE_API_PRIVATE_H
-#define MOD_PSU_DEVICE_API_PRIVATE_H
-
-#include <mod_psu.h>
-
-/* Module API implementation */
-extern const struct mod_psu_device_api __mod_psu_device_api;
-
-#endif /* MOD_PSU_DEVICE_API_PRIVATE_H */
