--- module/psu/src/mod_psu_private.h
@@ -1,16 +1,0 @@
-/*
- * Arm SCP/MCP Software
- * Copyright (c) 2017-2019, Arm Limited and Contributors. All rights reserved.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-#ifndef MOD_PSU_PRIVATE_H
-#define MOD_PSU_PRIVATE_H
-
-#include <fwk_id.h>
-#include <mod_psu_device_api_private.h>
-#include <mod_psu_event_private.h>
-#include <mod_psu_module_private.h>
-
-#endif /* MOD_PSU_PRIVATE_H */
